Fix Safari render/paint loop
Fix Safari render/paint loop

**What:**

As identified in [this bug
report](https://bugs.webkit.org/show_bug.cgi?id=194332) with additional
comments clarifying my findings hopefully soon to appear, there is a bug
in Safari that starts and endless cycle of render/paints when all the
following conditions are met:

1. a `transition` animation is defined that is not instantaneous
2. In the destination of the transition, an animating property contains
   a color value with:
  1. an alpha channel that is *exclusively* between 0 and 1
  2. at least one color channel which is *exclusively* between 0 and 255

These conditions represent my current best guess at the exact
requirements to encounter this bug.

You can explore the bug in this fiddle I made while trying to pinpoint
the minimum reproduction: https://jsfiddle.net/0ph9xkrt/7/

The bug described above affects this template only on the `.post-card`
class. All other animations (that I am aware of and have tested)
involving an element with an `rgba` color in one of its properties do
not directly animate the property containing the color, and are thus not
triggering this bug.

**How:**

There are three ways we can avoid the bug for the hover animation of
`.post-card`s in this template. One, we could just alter the shadow's
color to be a pure black (nullifying condition 2.2) and then compensate
the darker color with a slightly more transparent alpha channel. Two, we
could use the actual design colors (which actually contain a hint of
color) but skip animating the distance/blur of the shadow which means we
do not have to animate a property containing the color (nullifying
condition 2). Or three, we can use pseudo-selectors to define `:before`
and `:after` with each representing one state for the box shadow, then
animate only the opacity between the two (also nullifying condition 2).

This third option was the choice here since it will keep the design of
the site as is, visually, without compromising the color of the shadow
or the changable size of the shadow. It does have one drawback that the
other two options do not: the css is much more complex, as you'll see in
the diff. Not only do we need to add all the different states for two
new pseudo-selectors, but we also need to ensure that the links under
them are set to have a higher z-index so they remain clickable.