Fix multi-byline
Note:
The original method of applying a "hovered" class is actually better,
because when controlling in js, you can debounce. But that js logic
didn't appear to make the transition to this template in the first
place, so this is a starting point at least.

How:
**Fix typo in `post.hbs`**
This typo would force the single byline every time. Now, multiple
bylines are possible to render.

**Fix template for `byline-multiple.hbs`**
There were lots of things in here that did not get converted to Ember's
flavor of HBS. I am not aware of a `#foreach` block, but it seemed
broken, so I replaced one with an `#each` block and just completely
removed the other one, since it would presumably merely print the
authors' names, which are available on the cards.